Dynamic and Static Versions Update
- PlayerMirror Prefab
- MirrorPlayer.cs
- PlayerReference.cs
- SpaceShipReference.cs

The point of this update is to establish a connection between the static and dynamic versions of the space ships where one ship moves and one ship does not.  One player moves and the other simply mirrors its position and rotation.  This complicated solution solves the issue of having a dynamic rigidbody inside of another dynamic rigidbody.  In this solution the player inside of the static ship has a rigidbody and the player inside of the dynamic ship does not.  Therefore each version only has one dynamic rigidbody in their hierarchy.

Currently Space Ship Movement has been added and appears to work perfectly.  This has not been networked yet.